Tighten type annotations for stricter mypy
Mechanical type-hint hardening across the DSP, protocol-encoder, waterfall
and I/Q-utility modules so mypy stays clean under the stricter settings:
wrap bare numpy scalar types in np.dtype(...), annotate buffers that start
as None or empty arrays, use numpy.typing.DTypeLike for dtype parameters,
cast mixed int/float returns to float, and avoid rebinding a list variable
to an ndarray. No behaviour change.

+++ src/sdr_module/dsp/frequency_lock.py
@@ -254,7 +254,9 @@ class FrequencyLocker:
             signal_freq = center_freq + freq_offset
 
             # Estimate bandwidth (3 dB down from peak)
-            bandwidth = self._estimate_bandwidth(spectrum_db, peak_idx, freq_per_bin)
+            bandwidth = self._estimate_bandwidth(
+                spectrum_db, int(peak_idx), freq_per_bin
+            )
 
             # Calculate SNR
             snr_db = peak_power - self._noise_floor_db
@@ -440,7 +442,7 @@ class FrequencyLocker:
             stability_quality = 0.5
 
         # Combined quality
-        return 0.6 * snr_quality + 0.4 * stability_quality
+        return float(0.6 * snr_quality + 0.4 * stability_quality)
 
     def lock_to_frequency(self, frequency_hz: float) -> None:
         """
